use common inlined code for low-level dot products
